Fixed gpus parameter to allow for fractional values.

+++ config/default_args.py
@@ -76,13 +76,13 @@ def add_default_args(parser):
         "--cpus_for_driver", type=int, default=1, help="Number of CPUs used by the driver"
     )
     parser.add_argument(
-        "--gpus_for_driver", type=int, default=0, help="Number of GPUs used by the driver"
+        "--gpus_for_driver", type=float, default=0, help="Number of GPUs used by the driver"
     )
     parser.add_argument(
         "--cpus_per_worker", type=int, default=1, help="Number of CPUs used by one worker"
     )
     parser.add_argument(
-        "--gpus_per_worker", type=int, default=0, help="Number of GPUs used by one worker"
+        "--gpus_per_worker", type=float, default=0, help="Number of GPUs used by one worker"
     )
 
     parser.add_argument(
